refactor(chain): extract hardfork application into database_hardfork.cpp
database.cpp is the single longest-to-compile translation unit in the project
(~8,000 lines, one object file, cannot be parallelized). This moves the
self-contained hardfork cluster into its own TU so it compiles in parallel with
the rest of database.cpp and edits to hardfork logic no longer force a recompile
of the entire file.

Moved verbatim (byte-for-byte) from database.cpp into database_hardfork.cpp:
  init_hardforks, reset_virtual_schedule_time, process_hardforks,
  has_hardfork, set_hardfork, apply_hardfork
(the contiguous block that was lines 7012-7989). database.cpp shrinks by ~975
lines (7991 -> 7016); the new file is ~1,020 lines.

No behavior change:
- Method bodies are identical; only their file location changed.
- The cluster touches only database members declared in database.hpp
  (_hardfork_times, _hardfork_versions, _log_hardforks), so a separate TU can
  define them. CHAIN_HARDFORK_*/HARDFORK_* constants come from the generated
  hardfork.hpp, which database.hpp includes, so the new file (which includes
  database.hpp first) sees them.
- The two VIRTUAL_SCHEDULE_LAP_LENGTH* macros are still used elsewhere in
  database.cpp (lines ~3348, ~3817), so they remain defined there; the new file
  gets its own copy because reset_virtual_schedule_time needs them. Kept in sync
  by a comment in both places.
- to256() is a small inline helper already duplicated per-TU (database.cpp and
  chain_evaluator.cpp both define it); the new TU gets the same local copy.
- Added database_hardfork.cpp to both the SHARED and STATIC source lists in
  libraries/chain/CMakeLists.txt. The existing
  add_dependencies(graphene_chain ... build_hardfork_hpp) already covers the new
  TU's need for the generated header.

Verification (full tree build not possible in my env - no Boost/submodules):
- Symbol accounting: 163 database:: definitions/references before == 163 after
  (db + hardfork), none lost, none added.
- The original 7012-7989 block appears verbatim in the new file.
- Both files are brace-balanced; namespace nesting is closed once.
- The added scaffolding (the two macros, the to256 helper, u256 usage, namespace)
  passes g++ -std=c++14 -fsyntax-only against minimal stubs.
